Rename the nullifier test module
