+++ zebra-state/src/service/check/tests.rs
@@ -1,4 +1,4 @@
 //! Tests for state contextual validation checks.
 
-mod prop;
+mod nullifier;
 mod vectors;
